rename directory quiz -> mqtt
